Upload of new 5km grid shapefile
Upload of fivekm_grid_polys_shore_lamb.shp, which is based on the master_5km_grid_tmer.shp, but with a different coordinate system and polygons that span the shoreline are clipped to that shoreline. Polygon ID attribute values are identical between the two 5km grids, so joining is unaffected.