Shape- and wkt-files for rotated thematic maps in R
Shapefiles, in a custom Rectified Skew Orthomorphic coordinate reference system (CRS) that can be used for generating thematic maps in R that are rotated by 21 or 30 degrees. Each shapefile is a polygon defined by the spatial extent of the 5km grid used for generating heatmaps. The CRS for these shapefiles can be read in to R using either the shapefiles themselves (corresponding .prj files), or by reading in the corresponding .wkt files. Refer to README_map_rotation_CRSs.txt for more details.