Parameters for creating 5-panel zoomed maps
Details that people can use to generate 5-panel maps of the US West Coast, based on information from ArcGIS ArcMap. The left panel (A) is the entire west coast, and panels B-E correspond to 4-equal sized regions that are enlarged to show more detail. The information should be readily transferable to other environments, including R and QGIS. Make sure to look at the README worksheet first!